Refactor yaml utilities into separate directory
