+++ clowder/clowder/yaml/validation/fork_validation.py
@@ -1,0 +1,43 @@
+# -*- coding: utf-8 -*-
+"""clowder.yaml parsing and validation functionality
+
+.. codeauthor:: Joe Decapo <joe@polka.cat>
+
+"""
+
+from __future__ import print_function
+
+import clowder.util.formatting as fmt
+from clowder.error.clowder_error import ClowderError
+from clowder.yaml.util import (
+    dict_contains_value,
+    validate_type
+)
+
+
+def validate_yaml_fork(fork, yaml_file):
+    """Validate fork in clowder loaded from yaml file
+
+    :param dict fork: Parsed YAML python object for fork
+    :param str yaml_file: Path to yaml file
+    :return:
+    :raise ClowderError:
+    """
+
+    validate_type(fork, 'fork', dict, 'dict', yaml_file)
+
+    if not fork:
+        error = fmt.missing_entries_error('fork', yaml_file)
+        raise ClowderError(error)
+
+    dict_contains_value(fork, 'fork', 'name', yaml_file)
+    validate_type(fork['name'], 'name', str, 'str', yaml_file)
+    del fork['name']
+
+    dict_contains_value(fork, 'fork', 'remote', yaml_file)
+    validate_type(fork['remote'], 'remote', str, 'str', yaml_file)
+    del fork['remote']
+
+    if fork:
+        error = fmt.unknown_entry_error('fork', fork, yaml_file)
+        raise ClowderError(error)
